Move sim results conversion and set start timestamp
